Move from ActiveModelSerializers::Model to ActiveModel::Serializer::Model
The `ActiveModelSerializers::Model` is the only public API under this namespace.

To keep consistency move to `ActiveModel::Serializer::Model`.
