--- test/active_model_serializers/model_test.rb
@@ -1,9 +1,0 @@
-require 'test_helper'
-
-class ActiveModelSerializers::ModelTest < Minitest::Test
-  include ActiveModel::Serializer::Lint::Tests
-
-  def setup
-    @resource = ActiveModelSerializers::Model.new
-  end
-end
